requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
pydantic
python-dotenv